functions/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-vision
google-cloud-firestore
firebase-admin
